Add Android(Suite|Detail|Graph)Parser classes
