+++ backend/src/org/commcare/xml/AndroidDetailParser.java
@@ -1,0 +1,17 @@
+package org.commcare.xml;
+
+import org.kxml2.io.KXmlParser;
+
+/**
+ * Created by jschweers on 1/28/2016.
+ */
+public class AndroidDetailParser extends DetailParser {
+    public AndroidDetailParser(KXmlParser parser) {
+        super(parser);
+    }
+
+    @Override
+    protected GraphParser getGraphParser() {
+        return new AndroidGraphParser(parser);
+    }
+}
